Extracted common base type for old and new AbstractContentASsistParser
